Start station implemented
- Created start station
- Implemented start station into map generation
- Fixed Ground Collider Issues (Tile required BoxCollider for unknown reason)

+++ Assets/Scripts/Terrain/Generation/MapManager.cs
@@ -11,7 +11,7 @@ namespace Uncooked.Terrain.Generation
     public class MapManager : MonoBehaviour
     {
         [Header("Generation")] [SerializeField] private int seed = 0;
-        [SerializeField] private float noiseScale = 10;
+        [SerializeField] [Min(1)] private float noiseScale = 10;
         [SerializeField] private Vector2 noiseOffset;
         [Range(1, 1000)] [SerializeField] private int mapLength = 25, mapWidth = 19;
 
@@ -21,6 +21,15 @@ namespace Uncooked.Terrain.Generation
         private System.Random rng;
         [Header("Parents")] [SerializeField] private Transform floorParent;
         [SerializeField] private Transform obstacleParent, railParent;
+
+        [Header("Station")]
+        [SerializeField] private GameObject stationPrefab;
+        [SerializeField] private Tile stationBridge;
+        [SerializeField] private Vector2Int stationSize = 5 * Vector2Int.one;
+
+        private Vector3Int stationPos;
+
+        public Vector3 StartPoint => stationPos;
 
         public void Start()
         {
@@ -35,13 +44,20 @@ namespace Uncooked.Terrain.Generation
         /// </summary>
         public void GenerateMap()
         {
+            if (!stationPrefab) throw new System.Exception("No stationPrefab given");
+
             System.Action<Object> destroyType = DestroyImmediate;
             if (Application.isPlaying) destroyType = Destroy;
             foreach (Transform t in transform.Cast<Transform>().ToList()) destroyType(t.gameObject);
 
             rng = new System.Random(seed);
-            HUDManager.instance.UpdateSeedText(seed.ToString());
-            var heightMap = GenerateHeightMap(noiseScale);
+            if (Application.isPlaying) HUDManager.instance.UpdateSeedText(seed.ToString());
+            var heightMap = GenerateHeightMap();
+
+            // Station info
+            int stationHalfX = stationSize.x / 2, stationHalfZ = stationSize.y / 2;
+            stationPos = new Vector3Int(stationHalfX, 0, rng.Next(stationHalfZ, mapWidth - stationHalfZ));
+            BoundsInt stationBounds = new BoundsInt(stationPos - new Vector3Int(stationHalfX, 0, stationHalfZ), new Vector3Int(stationSize.x, 3, stationSize.y));
 
             #region Map Floor
             floorParent = new GameObject("Ground").transform;
@@ -55,9 +71,16 @@ namespace Uncooked.Terrain.Generation
                 for (int z = 0; z < mapWidth; z++)
                 {
                     // Gets tile type based on height map
-                    Tile tile = groundBiome.GetTile(heightMap[x, z]);
+                    var tile = groundBiome.GetTile(heightMap[x, z]);
                     var obj = Instantiate(tile, new Vector3(x, transform.position.y, z), Quaternion.identity, rowParent);
                     obj.name = obj.name.Substring(0, obj.name.Length - 7) + " " + z;
+
+                    // Adds bridge to liquid tiles under the station
+                    if (obj is LiquidTile liquid && stationBounds.Contains(new Vector3Int(x, (int)transform.position.y, z)))
+                    {
+                        Instantiate(stationBridge, liquid.transform.position, liquid.transform.rotation, liquid.transform.parent);
+                        liquid.GetComponent<BoxCollider>().enabled = false;
+                    }
                 }
             }
 
@@ -72,6 +95,9 @@ namespace Uncooked.Terrain.Generation
             obstacleParent = new GameObject("Obstacles").transform;
             obstacleParent.parent = transform;
 
+            // Station placement
+            Instantiate(stationPrefab, stationPos, Quaternion.identity, obstacleParent);
+
             for (int x = 0; x < mapLength; x++)
             {
                 var rowParent = new GameObject("Row " + x).transform;
@@ -79,6 +105,8 @@ namespace Uncooked.Terrain.Generation
 
                 for (int z = 0; z < mapWidth; z++)
                 {
+                    if (stationBounds.Contains(new Vector3Int(x, (int)transform.position.y + 1, z))) continue;
+
                     // Gets tile type based on height map
                     Tile tile = obstacleBiome.GetTile(heightMap[x, z]);
                     if (tile == null) continue;
@@ -92,11 +120,11 @@ namespace Uncooked.Terrain.Generation
         }
 
         /// <summary>
-        /// Generates 2D perlin noise based on Generation variables
+        /// Generates 2D perlin noise based on generation variables
         /// </summary>
         /// <param name="noiseScale"></param>
         /// <returns></returns>
-        private float[,] GenerateHeightMap(float noiseScale)
+        private float[,] GenerateHeightMap()
         {
             float[,] heightMap = new float[mapLength, mapWidth];
             Vector2 offset = new Vector2(rng.Next(-100000, 100000), rng.Next(-100000, 100000));
@@ -134,6 +162,12 @@ namespace Uncooked.Terrain.Generation
         {
             obstacleParent.gameObject.SetActive(enabled);
             floorParent.gameObject.SetActive(enabled);
+        }
+
+        void OnValidate()
+        {
+            if (stationSize[0] < 1) stationSize[0] = 1;
+            if (stationSize[1] < 1) stationSize[1] = 1;
         }
     }
 }
